// convex/_generated/api.d.ts
// This file will be auto-generated when you run `npx convex dev` or `npx convex codegen`
// For now, this is a placeholder to allow the code to compile

export const api: {
  teams: {
    create: any;
    get: any;
    getUserTeams: any;
  };
  teamMembers: {
    add: any;
    getByTeam: any;
    isMember: any;
  };
  wishlists: {
    upsert: any;
    get: any;
    getByTeam: any;
  };
  assignments: {
    create: any;
    get: any;
    hasAssignments: any;
  };
  users: {
    get: any;
    upsert: any;
  };
  sessions: {
    get: any;
    set: any;
    destroy: any;
    clearExpired: any;
  };
};